include/essentials: Define `UK_NACTUAL_ARGS` macro
Implement the actual argument count of `UK_NARGS`. This can be used
in the context of a system call registration, but it can be used to
get the actual number of arguments (i.e. excluding types).

The `7`'s represent invalid numbers, as, at the moment, there is no
macro with the `7` suffix that may end up being executed by mistake.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>