lib/uksched: Do not save/restore ECTX on coop thread switching
Since we only have cooperative scheduling at the moment, saving/restoring
ECTX when thread switching is not necessary, since the ABI treats these
registers as scratch registers and thread switching essentially
happens through actual explicit function calls.

On the other hand, this is obviously not the case with preemptive
scheduling since thread switching can happen out of the blue,
typically on a timer IRQ, so then ECTX saving/restoring is definetely
necessary.

But, again, we are only doing cooperative scheduling at the moment and
this operation adds unnecessary overhead so throw it out.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>