lib/ramfs: Explicitly mention that there are no public symbols
`exportsyms.uk` of RAMFS had a `none` dummy string so that no symbols
are exported, otherwise the `objcopy --keep-global-symbols=` would have
failed due to an empty file being given as argument.
However, this is inconsistent with other such files which just had
`# No public symbols` which is just as effective. Thus make it
consistent and do the same here.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>