lib/syscall_shim: Use regular function regs for execenv arg mappings
Use the registers that are defined to be used in the calling convention
of regular functions instead of syscalls when mapping execenv registers.
This is highly important for x86 as, unlike ARM whose mapping is the
same between function/syscall, its 4th argument is r10 instead of rcx.
So this actually also fixes a bug waiting to happen.

At the moment of writing this, the only syscall with more than 3 args
that uses the execenv registration is clone and so its `child_tid`
argument would end up being wrong. Luckily, we haven't used that one
so far.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>